fix/chat_feature_synthesis_to_service
