+++ src/app/new-chat/new-chat.component.html
@@ -1,0 +1,7 @@
+<app-sidebar></app-sidebar>
+<script src="https://unpkg.com/typewriter-effect@latest/dist/core.js"></script>
+<div class="card-container">
+
+</div>
+
+
